fix rare bug with escape sequences
Strings with an even number of consecutive escape characters (backslashes), could potentially throw the script off in some scenarios like this one:

user_pref("trollface", "ILikeBackslashes\\"blahblah" );

To the Firefox parser, that would be a syntax error: the quotes-within-quotes would be interpreted as the end of the string. The script wouldn't know this, though. With this fix, it now makes sure there is an odd number of backslashes preceding quotes or double quotes within strings, and so interprets them the same way Firefox does.

Did I mention I'm a regex junkie?